requirements.txt

mysql-connector-python==8.0.28
apscheduler==3.9.1
psutil==5.9.0
